pyproject.toml
[project]
name = "todotasks"
version = "0.1.0"
description = "TodoList FastAPI application"
authors = [
    {name = "klim",email = "thrandi230807@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "fastapi-cli (>=0.0.8,<0.0.9)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
